slamd-96: Renamed discovery_conductor.py -> experiment_conductor.py
